[modified] mainpage 컴포넌트화 작업 - footer 재사용 가능
