+++ FrontEnd/package.json
@@ -16,7 +16,6 @@
     "@types/react-dom": "^18.2.1",
     "axios": "^1.4.0",
     "moment": "^2.29.4",
-    "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.1",
     "react-scripts": "5.0.1",
